B56_upd2.cpp

+++ src/alg.cpp
@@ -40,7 +40,7 @@ uint64_t nextPrime(uint64_t value) {
 }
 
 uint64_t sumPrime(uint64_t hbound) {
-	int s = 0;
+	uint64_t s = 0;
 	for (uint64_t i = 2; i < hbound; i++) {
 		if (checkPrime(i)) {
 			s+=i;
